+++ src/main/java/pl/coderslab/charity/service/UserService.java
@@ -13,6 +13,5 @@ public interface UserService {
 
     boolean userExist(UserDto userDto);
 
-    Integer numberOfUsers();
-
+    Long numberOfUsers();
 }
